+++ src/app/globals.css
@@ -38,16 +38,16 @@ body {
     --chart-6: hsl(var(--opposite-hue) 50% 90%);
 
     /* Task Color Palette */
-    --task-color-1: 4 90% 58%;   /* Tomato */
-    --task-color-2: 36 90% 58%;  /* Tangerine */
-    --task-color-3: 54 90% 58%;  /* Citron */
-    --task-color-4: 90 90% 58%;  /* Lime */
-    --task-color-5: 150 90% 58%; /* Mint */
-    --task-color-6: 190 90% 58%; /* Sky */
-    --task-color-7: 250 90% 65%; /* Lavender */
-    --task-color-8: 280 90% 65%; /* Plum */
-    --task-color-9: 340 90% 65%; /* Rose */
-    --task-color-10: 240 10% 40%;/* Graphite */
+    --task-color-1: 4 75% 90%;   /* Tomato */
+    --task-color-2: 36 75% 90%;  /* Tangerine */
+    --task-color-3: 54 75% 90%;  /* Citron */
+    --task-color-4: 90 75% 90%;  /* Lime */
+    --task-color-5: 150 75% 90%; /* Mint */
+    --task-color-6: 190 75% 90%; /* Sky */
+    --task-color-7: 250 75% 90%; /* Lavender */
+    --task-color-8: 280 75% 90%; /* Plum */
+    --task-color-9: 340 75% 90%; /* Rose */
+    --task-color-10: 240 75% 90%;/* Graphite */
 
     --radius: 0.5rem;
     --sidebar-background: 0 0% 98%;
